fix(gauge): resolve ESLint rule violations and ref access during render
